backend/requirements.txt
django==5.1.2
celery==5.4.0
djangorestframework==3.15.2
redis==5.2.0
gunicorn==23.0.0
python-dotenv==1.0.1